// packages/frontend/e2e-smoke-tests/interfaces/dashboardID-interfaces.ts
export enum EDashboardID {
    chart = '1783466d-aaf6-4662-aeaf-84e289ab8720',
    chartFormula = 'e517a563-1f85-4381-8a86-7206ce209503',
    chartLevels = '0b04e78e-6b8d-4401-ae4c-0fdf037acae5',
    chartScheme = '46bfadda-13c4-448a-8274-2e7099ec8304',
    chartSettings = '71fba67f-70a2-439d-aa72-009989d62601',
    grid = 'd0724f18-59e1-452b-ab15-0aef0a5ed1b9',
    gridFeatures = '067799ad-67fc-4829-9b8b-ca4ca8212f3c',
    panels = '5c64038b-c43e-4e82-9029-0187df231fa7',
    panelsLayouts = '00ce3107-4a2f-46cf-a629-112c24f7645b',
    panelsWithType = '3d24cc28-4ac7-4649-89c8-d727465f3966',
    panelsWithoutType = 'e0afb783-f95c-43ab-9e96-10bcfd560aa1',
    tableFeatures = '36e2e90a-f886-4bd2-a415-28b9f4a5db56',
}